completed all api

+++ convex/users.ts
@@ -1,7 +1,5 @@
 import {ConvexError, v} from "convex/values";
 import {MutationCtx, QueryCtx, internalMutation, mutation, query} from "./_generated/server";
-import {getCourse} from "./courses";
-import {getProblem} from "./problems";
 import {activity} from "./schema";
 
 export async function getUser(ctx: QueryCtx | MutationCtx, tokenIdentifier: string) {
@@ -11,10 +9,99 @@ export async function getUser(ctx: QueryCtx | MutationCtx, tokenIdentifier: stri
 		.first();
 
 	if (!user) {
-		throw new ConvexError("expected user to be defined");
+		return null;
 	}
 
 	return user;
+}
+
+export async function deleteUserById(ctx: MutationCtx, userId: string) {
+	const user = await getUser(ctx, userId);
+	if (!user) {
+		return;
+	}
+	// Delete the user
+	await ctx.db.delete(user._id);
+
+	// Delete the role
+	const role = await ctx.db
+		.query("role")
+		.withIndex("by_userId", q => q.eq("userId", userId))
+		.first();
+	if (role) {
+		await ctx.db.delete(role._id);
+	}
+
+	// Delete the activity
+	const activity = await ctx.db
+		.query("activities")
+		.withIndex("by_userId", q => q.eq("userId", userId))
+		.first();
+	if (activity) {
+		await ctx.db.delete(activity._id);
+	}
+	// Delete the userCourse
+	const userCourse = await ctx.db
+		.query("userCourses")
+		.withIndex("by_userId", q => q.eq("userId", userId))
+		.collect();
+	if (userCourse) {
+		await Promise.all(
+			userCourse.map(async course => {
+				await ctx.db.delete(course._id);
+			}),
+		);
+	}
+	// Delete the favoriteProblems
+	const favoriteProblems = await ctx.db
+		.query("favoriteProblems")
+		.withIndex("by_userId", q => q.eq("userId", userId))
+		.collect();
+
+	if (favoriteProblems) {
+		await Promise.all(
+			favoriteProblems.map(async problem => {
+				await ctx.db.delete(problem._id);
+			}),
+		);
+	}
+
+	// Delete the notifiesToAdmin
+	const notifiesToAdmin = await ctx.db
+		.query("notifiesToAdmin")
+		.withIndex("by_userId", q => q.eq("userId", userId))
+		.collect();
+	if (notifiesToAdmin) {
+		await Promise.all(
+			notifiesToAdmin.map(async notify => {
+				await ctx.db.delete(notify._id);
+			}),
+		);
+	}
+	// Delete the notifiesToUser
+	const notifiesToUser = await ctx.db
+		.query("notifiesToUser")
+		.withIndex("by_userId", q => q.eq("userId", userId))
+		.collect();
+	if (notifiesToUser) {
+		await Promise.all(
+			notifiesToUser.map(async notify => {
+				await ctx.db.delete(notify._id);
+			}),
+		);
+	}
+	// Delete the codeOfUserInLesson
+	const codeOfUserInLesson = await ctx.db
+		.query("codeOfUserInLesson")
+		.withIndex("by_userId", q => q.eq("userId", userId))
+		.collect();
+	if (codeOfUserInLesson) {
+		await Promise.all(
+			codeOfUserInLesson.map(async code => {
+				await ctx.db.delete(code._id);
+			}),
+		);
+	}
 }
 
 export const createUser = internalMutation({
@@ -33,7 +120,9 @@ export const updateUser = internalMutation({
 	args: {userId: v.string(), name: v.string(), image: v.string()},
 	async handler(ctx, args) {
 		const user = await getUser(ctx, args.userId);
-
+		if (!user) {
+			return;
+		}
 		await ctx.db.patch(user._id, {
 			name: args.name,
 			image: args.image,
@@ -45,7 +134,9 @@ export const deleteUser = internalMutation({
 	args: {userId: v.string()},
 	async handler(ctx, args) {
 		const user = await getUser(ctx, args.userId);
-
+		if (!user) {
+			return;
+		}
 		await ctx.db.delete(user._id);
 	},
 });
@@ -53,19 +144,19 @@ export const deleteUser = internalMutation({
 //************************************** */
 
 export const getUserProfile = query({
-	args: {userId: v.id("users")},
+	args: {userId: v.string()},
 	async handler(ctx, args) {
-		const user = await ctx.db.get(args.userId);
-
-		return user;
+		return await getUser(ctx, args.userId);
 	},
 });
 
 export const updateUserIntroduction = mutation({
-	args: {userId: v.id("users"), introduce: v.string()},
+	args: {userId: v.string(), introduce: v.string()},
 	async handler(ctx, args) {
 		const user = await getUser(ctx, args.userId);
-
+		if (!user) {
+			throw new ConvexError("User not found");
+		}
 		await ctx.db.patch(user._id, {introduce: args.introduce});
 	},
 });
